Move captcha-related files to the core module
