+++ indico/modules/core/plugins.py
