+++ indico/web/client/js/react/components/Captcha.jsx
