+++ indico/web/client/js/react/components/Captcha.module.scss
